Log parsing improvements
Cache parsed log data, so we only need to parse changed/new logfiles.
Close matplotlib figures, as directed by the softwarez.
