+++ dreambooth/train_dreambooth.py
@@ -23,9 +23,13 @@ from packaging import version
 from tensorflow.python.framework.random_seed import set_seed as set_seed1
 from torch.cuda.profiler import profile
 from torch.utils.data import Dataset
+from torch.utils.tensorboard import SummaryWriter, FileWriter
 from transformers import AutoTokenizer
 
+
+
 try:
+    from extensions.sd_dreambooth_extension.helpers.log_parser import LogParser
     from extensions.sd_dreambooth_extension.dreambooth import xattention, shared
     from extensions.sd_dreambooth_extension.dreambooth.dataclasses.prompt_data import PromptData
     from extensions.sd_dreambooth_extension.dreambooth.dataclasses.train_result import TrainResult
@@ -40,7 +44,7 @@ try:
     from extensions.sd_dreambooth_extension.dreambooth.utils.model_utils import unload_system_models, \
     import_model_class_from_model_name_or_path, disable_safe_unpickle, enable_safe_unpickle, xformerify, torch2ify
     from extensions.sd_dreambooth_extension.dreambooth.utils.text_utils import encode_hidden_state
-    from extensions.sd_dreambooth_extension.dreambooth.utils.utils import cleanup, parse_logs, printm
+    from extensions.sd_dreambooth_extension.dreambooth.utils.utils import cleanup, printm
     from extensions.sd_dreambooth_extension.dreambooth.webhook import send_training_update
     from extensions.sd_dreambooth_extension.dreambooth.xattention import optim_to
     from extensions.sd_dreambooth_extension.helpers.ema_model import EMAModel
@@ -50,6 +54,7 @@ try:
         TEXT_ENCODER_DEFAULT_TARGET_REPLACE, get_target_module
     from extensions.sd_dreambooth_extension.dreambooth.deis_velocity import get_velocity
 except:
+    from dreambooth.helpers.log_parser import LogParser
     from dreambooth.dreambooth import xattention, shared  # noqa
     from dreambooth.dreambooth.dataclasses.prompt_data import PromptData  # noqa
     from dreambooth.dreambooth.dataclasses.train_result import TrainResult  # noqa
@@ -63,7 +68,7 @@ except:
     from dreambooth.dreambooth.utils.image_utils import db_save_image, get_scheduler_class  # noqa
     from dreambooth.dreambooth.utils.model_utils import unload_system_models, import_model_class_from_model_name_or_path, disable_safe_unpickle, enable_safe_unpickle  # noqa
     from dreambooth.dreambooth.utils.text_utils import encode_hidden_state  # noqa
-    from dreambooth.dreambooth.utils.utils import cleanup, parse_logs, printm  # noqa
+    from dreambooth.dreambooth.utils.utils import cleanup, printm  # noqa
     from dreambooth.dreambooth.webhook import send_training_update  # noqa
     from dreambooth.dreambooth.xattention import optim_to  # noqa
     from dreambooth.helpers.ema_model import EMAModel  # noqa
@@ -138,6 +143,7 @@ def main(use_txt2img: bool = True) -> TrainResult:
     """
     args = shared.db_model_config
     logging_dir = Path(args.model_dir, "logging")
+    log_parser = LogParser()
 
     result = TrainResult
     result.config = args
@@ -176,6 +182,23 @@ def main(use_txt2img: bool = True) -> TrainResult:
                 project_dir=logging_dir,
                 cpu=shared.force_cpu
             )
+
+            run_name = "dreambooth.events"
+            max_log_size = 250*1024  # specify the maximum log size
+
+            # create a SummaryWriter object with max_queue and flush_secs arguments
+            writer = SummaryWriter(
+                log_dir=logging_dir,
+                filename_suffix=run_name,
+                max_queue=max_log_size,
+                flush_secs=30  # how often to flush the pending events to disk (in seconds)
+            )
+
+            # get the tracker object for TensorBoard
+            tracker = Accelerator.get_tracker("tensorboard")
+
+            # set the SummaryWriter object as the tracker's writer
+            tracker.writer = FileWriter(writer.get_logdir())
         except Exception as e:
             if "AcceleratorState" in str(e):
                 msg = "Change in precision detected, please restart the webUI entirely to use new precision."
@@ -853,7 +876,7 @@ def main(use_txt2img: bool = True) -> TrainResult:
                         del generator
                     try:
                         printm("Parse logs.")
-                        log_images, log_names = parse_logs(model_name=args.model_name)
+                        log_images, log_names = log_parser.parse_logs(model_name=args.model_name)
                         pbar.update()
                         for log_image in log_images:
                             last_samples.append(log_image)
